Refactor clean up
Refactored application testing files to support adding a new folder.

Testing still causes an issue until this initial 5 increments of the id are passed; not sure of the reasoning here but the application does return successfully.

This will need to be tested in a production enveironment to identify if it is an issue or if this is specific to the testing environment caused by the truncation of the databased before and after each test is executed.

Next is to refactor to create routes for both folders and note endpoints.